+++ Project/Code/app/Http/Controllers/LocalizationController.php
@@ -1,9 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-use App;
+use Illuminate\Support\Facades\App;
 use App\Models\ResearchGroup;
 use App\Models\ResearchProject;
 use Illuminate\Support\Facades\Config;
@@ -18,10 +16,12 @@ class LocalizationController extends Controller
        // return view('welcome');
     }
     public function switchLang($lang)
-    {
-        if (array_key_exists($lang, Config::get('languages'))) {
-            Session::put('applocale', $lang);
-        }
-        return redirect()->back();
+{
+    // ตรวจสอบว่า $lang อยู่ใน array ของภาษาที่รองรับ
+    if (array_key_exists($lang, Config::get('languages'))) {
+        Session::put('applocale', $lang); // เก็บภาษาที่เลือกใน session
+        App::setLocale($lang); // เปลี่ยนภาษาใน app
     }
+    return redirect()->back(); // กลับไปหน้าก่อนหน้า
+}
 }