+++ Project/Code/config/app.php
@@ -236,6 +236,8 @@ return [
         'Html' => Collective\Html\HtmlFacade::class,
         'DataTables' => Yajra\DataTables\Facades\DataTables::class,
         'PDF' => Barryvdh\DomPDF\Facade::class,
+        'locale' => 'en',
+        'fallback_locale' => 'en'
     ],
 
 ];