+++ Project/Code/resources/views/dashboards/admins/layouts/admin-dash-layout.blade.php
@@ -205,7 +205,6 @@ scratch. This page gets rid of all links and provides the needed markup only.
                     </a>
                 </li>
                 @endcan
-                
                 @can('departments-list')
                 <li class="nav-item">
                     <a class="nav-link" href="{{ route('departments.index')}}">